refactor(use-cases): extract service platforms into section component
