+++ src/components/use-cases/index.ts
@@ -1,0 +1,1 @@
+export { default as ServicePlatformsSection } from "./ServicePlatformsSection";
